requirements.txt
aiohttp==3.7.4.post0
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
# discord-py-slash-command==4.2.1
discord-py-interactions==5.11.0
# discord.py==2.3.2
idna==3.6
multidict==6.0.5
pycparser==2.21
PyNaCl==1.5.0
python-dotenv==1.0.1
requests==2.31.0
six==1.16.0
spotipy==2.23.0
typing-extensions==4.9.0
Unidecode==1.3.8
urllib3==2.2.0
websocket-client==1.7.0
yarl==1.9.4
youtube-dl==2021.12.17
ytmusicapi==1.5.2
ffmpeg
